Custom Hydrophone and NC file read capacities
Add a custom or generic hydrophone class which allows users to have the basic capabilities to work with with the pypam package with less proscribed processes. Edits to init and add the 'custom' file.

Additionally added the capability to read calibration information from nc files. Data should be in 'frequency' and 'sensitivity' variables

print('NC file found looking for frequency and sensitivity variables')
            data = Dataset(self.calibration_file, 'r')  # 'r' means read mode

            # Frequency and sensitivity response
            freq = data.variables['frequency'][:]
            sendb = data.variables['sensitivity'][:]
            df = pd.DataFrame({'frequency': freq, 'Column2': sendb})
            df.columns = ['frequency', val]